refactor: refactore the app folders and files for increasing AI usability
