+++ app/api/endpoints/user/auth.py
@@ -7,10 +7,10 @@ from datetime import timedelta
 from sqlalchemy.orm import Session
 
 # import
-from app.schemas.user import User, UserLogin, Token
+from app.schemas.user_schema import User, UserLogin, Token
 from app.core.dependencies import get_db
 from app.core.settings import ACCESS_TOKEN_EXPIRE_MINUTES, REFRESH_TOKEN_EXPIRE_DAYS
-from app.api.endpoints.user import functions as user_functions
+from app.api.endpoints.user import user_function as user_functions
 
 
 auth_module = APIRouter()
